+++ src/MyCerts/UI/CategoryController.php
@@ -2,18 +2,16 @@
 
 namespace MyCerts\UI;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use MyCerts\Domain\Model\Category;
-use MyCerts\Domain\Roles;
 
 class CategoryController extends BaseController
 {
     public function list()
     {
-        $nonAdminRestriction = Auth::user()->isAdmin() ? [] : ['company_id'=>Auth::user()->company_id];
+        $nonAdminRestriction = Auth::user()->isAdmin() ? [] : ['company_id' => Auth::user()->company_id];
 
         return response()->json(Category::where($nonAdminRestriction)->paginate(self::DEFAULT_PAGINATION_LENGHT));
     }
@@ -21,14 +19,18 @@ class CategoryController extends BaseController
     public function create(Request $request)
     {
         $this->validate($request, [
-            'name'         => 'required',
+            'name'        => 'required|string',
+            'description' => 'string',
+            'icon'        => 'string',
         ]);
 
         $company = $this->retrieveCompany($request);
 
         $entity = new Category(array_filter([
-            'company_id' => $company->id,
-            'name'       => $request->get('name'),
+            'company_id'  => $company->id,
+            'name'        => $request->json('name'),
+            'description' => $request->json('description'),
+            'icon'        => $request->json('icon'),
         ]));
 
         $entity->save();
@@ -39,13 +41,17 @@ class CategoryController extends BaseController
     public function patch(string $id, Request $request)
     {
         $this->validate($request, [
-            'name'         => 'required',
+            'name'        => 'string',
+            'description' => 'string',
+            'icon'        => 'string',
         ]);
 
-        $company = $this->retrieveCompany($request);
+        $company  = $this->retrieveCompany($request);
         $category = Category::where(['id' => $id, 'company_id' => $company->id])->first();
         $category->fill(array_filter([
-            'name' => $request->json('name'),
+            'name'        => $request->json('name'),
+            'description' => $request->json('description'),
+            'icon'        => $request->json('icon'),
         ]));
         $category->save();
 
@@ -59,10 +65,12 @@ class CategoryController extends BaseController
 
     public function delete($id, Request $request)
     {
-        $company = $this->retrieveCompany($request);
+        $company  = $this->retrieveCompany($request);
         $category = Category::where(['id' => $id, 'company_id' => $company->id])->first();
-        $category->delete();
+        if ($category) {
+            $category->delete();
+        }
 
-        return response(null,Response::HTTP_NO_CONTENT);
+        return response(null, Response::HTTP_NO_CONTENT);
     }
 }
